+++ ChangeLog.md
@@ -2,4 +2,5 @@
 
 ## Unreleased changes
 
+* Added `Database.Franz.Reconnect`
 * Added `Database.Franz.URI`